requirements.txt
apache-airflow
requests
pandas
openmeteo_requests
psycopg2-binary
